movie_rent_backend/requirements.txt
Django==3.0.1
django-cors-headers==3.2.0
django-csvimport==2.13
django-import-export==2.0
djangorestframework==3.11.0
djoser==2.0.3